refactor(project): move template domain modules
