+++ crates/djls-project/src/sync.rs
@@ -8,7 +8,7 @@ use camino::Utf8PathBuf;
 use salsa::Setter;
 
 use crate::db::Db as ProjectDb;
-use crate::environment::templatetag_candidate_paths;
+use crate::templates::templatetag_candidate_paths;
 use crate::project::Project;
 use crate::resolve::SearchPaths;
 use crate::resolve::model_modules;
